+++ agentops/__init__.py
@@ -11,6 +11,7 @@ from .session import Session
 import threading
 from importlib.metadata import version as get_version
 from packaging import version
+from .llms import tracker
 
 try:
     from .partners.langchain_callback_handler import (
