+++ agentops/llms/providers/groq.py
@@ -1,31 +1,39 @@
-import inspect
 import pprint
 from typing import Optional
 
-from agentops.llms.instrumented_provider import InstrumentedProvider
-from agentops.time_travel import fetch_completion_override_from_time_travel_cache
-
-from ..event import ActionEvent, ErrorEvent, LLMEvent
-from ..session import Session
-from ..log_config import logger
-from ..helpers import check_call_stack_for_agent_id, get_ISO_time
-from ..singleton import singleton
+from .instrumented_provider import InstrumentedProvider
+from ...event import ErrorEvent, LLMEvent
+from ...session import Session
+from ...log_config import logger
+from agentops.helpers import get_ISO_time, check_call_stack_for_agent_id
+from ...singleton import singleton
 
 
 @singleton
-class OpenAiProvider(InstrumentedProvider):
+class GroqProvider(InstrumentedProvider):
     original_create = None
-    original_create_async = None
+    original_async_create = None
 
     def __init__(self, client):
         super().__init__(client)
-        self._provider_name = "OpenAI"
+        self.client = client
 
-    def handle_response(self, response, kwargs, init_timestamp, session: Optional[Session] = None) -> dict:
+    def override(self):
+        self._override_chat()
+        self._override_async_chat()
+
+    def undo_override(self):
+        if self.original_create is not None and self.original_async_create is not None:
+            from groq.resources.chat import completions
+
+            completions.Completions.create = self.original_create
+            completions.AsyncCompletions.create = self.original_create
+
+    def handle_response(self, response, kwargs, init_timestamp, session: Optional[Session] = None):
         """Handle responses for OpenAI versions >v1.0.0"""
-        from openai import AsyncStream, Stream
-        from openai.resources import AsyncCompletions
-        from openai.types.chat import ChatCompletionChunk
+        from groq import AsyncStream, Stream
+        from groq.resources.chat import AsyncCompletions
+        from groq.types.chat import ChatCompletionChunk
 
         llm_event = LLMEvent(init_timestamp=init_timestamp, params=kwargs)
         if session is not None:
@@ -113,7 +121,7 @@ class OpenAiProvider(InstrumentedProvider):
 
         # v1.0.0+ responses are objects
         try:
-            llm_event.returns = response
+            llm_event.returns = response.model_dump()
             llm_event.agent_id = check_call_stack_for_agent_id()
             llm_event.prompt = kwargs["messages"]
             llm_event.prompt_tokens = response.usage.prompt_tokens
@@ -135,102 +143,33 @@ class OpenAiProvider(InstrumentedProvider):
 
         return response
 
-    def override(self):
-        self._override_openai_v1_completion()
-        self._override_openai_v1_async_completion()
+    def _override_chat(self):
+        from groq.resources.chat import completions
 
-    def _override_openai_v1_completion(self):
-        from openai.resources.chat import completions
-        from openai.types.chat import ChatCompletion, ChatCompletionChunk
-
-        # Store the original method
         self.original_create = completions.Completions.create
 
         def patched_function(*args, **kwargs):
+            # Call the original function with its original arguments
             init_timestamp = get_ISO_time()
             session = kwargs.get("session", None)
             if "session" in kwargs.keys():
                 del kwargs["session"]
-
-            completion_override = fetch_completion_override_from_time_travel_cache(kwargs)
-            if completion_override:
-                result_model = None
-                pydantic_models = (ChatCompletion, ChatCompletionChunk)
-                for pydantic_model in pydantic_models:
-                    try:
-                        result_model = pydantic_model.model_validate_json(completion_override)
-                        break
-                    except Exception as e:
-                        pass
-
-                if result_model is None:
-                    logger.error(
-                        f"Time Travel: Pydantic validation failed for {pydantic_models} \n"
-                        f"Time Travel: Completion override was:\n"
-                        f"{pprint.pformat(completion_override)}"
-                    )
-                    return None
-                return self.handle_response(result_model, kwargs, init_timestamp, session=session)
-
-            # prompt_override = fetch_prompt_override_from_time_travel_cache(kwargs)
-            # if prompt_override:
-            #     kwargs["messages"] = prompt_override["messages"]
-
-            # Call the original function with its original arguments
             result = self.original_create(*args, **kwargs)
             return self.handle_response(result, kwargs, init_timestamp, session=session)
 
         # Override the original method with the patched one
         completions.Completions.create = patched_function
 
-    def _override_openai_v1_async_completion(self):
-        from openai.resources.chat import completions
-        from openai.types.chat import ChatCompletion, ChatCompletionChunk
+    def _override_async_chat(self):
+        from groq.resources.chat import completions
 
-        # Store the original method
-        self.original_create_async = completions.AsyncCompletions.create
+        self.original_async_create = completions.AsyncCompletions.create
 
         async def patched_function(*args, **kwargs):
-            init_timestamp = get_ISO_time()
-
-            session = kwargs.get("session", None)
-            if "session" in kwargs.keys():
-                del kwargs["session"]
-
-            completion_override = fetch_completion_override_from_time_travel_cache(kwargs)
-            if completion_override:
-                result_model = None
-                pydantic_models = (ChatCompletion, ChatCompletionChunk)
-                for pydantic_model in pydantic_models:
-                    try:
-                        result_model = pydantic_model.model_validate_json(completion_override)
-                        break
-                    except Exception as e:
-                        pass
-
-                if result_model is None:
-                    logger.error(
-                        f"Time Travel: Pydantic validation failed for {pydantic_models} \n"
-                        f"Time Travel: Completion override was:\n"
-                        f"{pprint.pformat(completion_override)}"
-                    )
-                    return None
-                return self.handle_response(result_model, kwargs, init_timestamp, session=session)
-
-            # prompt_override = fetch_prompt_override_from_time_travel_cache(kwargs)
-            # if prompt_override:
-            #     kwargs["messages"] = prompt_override["messages"]
-
             # Call the original function with its original arguments
-            result = await self.original_create_async(*args, **kwargs)
-            return self.handle_response(result, kwargs, init_timestamp, session=session)
+            init_timestamp = get_ISO_time()
+            result = await self.original_async_create(*args, **kwargs)
+            return self.handle_response(result, kwargs, init_timestamp)
 
         # Override the original method with the patched one
         completions.AsyncCompletions.create = patched_function
-
-    def undo_override(self):
-        if self.original_create is not None and self.original_create_async is not None:
-            from openai.resources.chat import completions
-
-            completions.AsyncCompletions.create = self.original_create_async
-            completions.Completions.create = self.original_create
