+++ agentops/llms/providers/mistral.py
@@ -3,9 +3,9 @@ import pprint
 import sys
 from typing import Optional
 
-from ..event import LLMEvent, ErrorEvent
-from ..session import Session
-from ..log_config import logger
+from ...event import LLMEvent, ErrorEvent
+from ...session import Session
+from ...log_config import logger
 from agentops.helpers import get_ISO_time, check_call_stack_for_agent_id
 from .instrumented_provider import InstrumentedProvider
